1. Refactor the write/read mode of the filesystem interface 2. Support extended object storage 3. Support multiple asynchronous write modes

+++ tskv/src/file_system/file/os/windows.rs
@@ -14,7 +14,7 @@ pub fn fd(file: &File) -> usize {
     file.as_raw_handle() as usize
 }
 
-pub fn pread(raw_handle: usize, pos: u64, len: usize, buf_ptr: u64) -> Result<usize> {
+pub fn pread(raw_handle: usize, pos: usize, len: usize, buf_ptr: u64) -> Result<usize> {
     let mut bytes: DWORD = 0;
     let mut ov = overlapped(pos);
     check_err(unsafe {
@@ -29,7 +29,7 @@ pub fn pread(raw_handle: usize, pos: u64, len: usize, buf_ptr: u64) -> Result<us
     Ok(usize::try_from(bytes).unwrap())
 }
 
-pub fn pwrite(raw_handle: usize, pos: u64, len: usize, buf_ptr: u64) -> Result<usize> {
+pub fn pwrite(raw_handle: usize, pos: usize, len: usize, buf_ptr: u64) -> Result<usize> {
     let mut bytes: DWORD = 0;
     let mut ov = overlapped(pos);
     check_err(unsafe {
@@ -44,15 +44,15 @@ pub fn pwrite(raw_handle: usize, pos: u64, len: usize, buf_ptr: u64) -> Result<u
     Ok(bytes as usize)
 }
 
-pub fn file_size(raw_handle: usize) -> Result<u64> {
+pub fn file_size(raw_handle: usize) -> Result<usize> {
     let mut info = MaybeUninit::<BY_HANDLE_FILE_INFORMATION>::zeroed();
     check_err(unsafe { GetFileInformationByHandle(raw_handle as RawHandle, info.as_mut_ptr()) })?;
     let info = unsafe { info.assume_init() };
     let len = u64::from(info.nFileSizeHigh) << 32 | u64::from(info.nFileSizeLow);
-    Ok(len)
+    Ok(len as usize)
 }
 
-fn overlapped(pos: u64) -> OVERLAPPED {
+fn overlapped(pos: usize) -> OVERLAPPED {
     unsafe {
         let mut r: OVERLAPPED = std::mem::zeroed();
         r.u.s_mut().Offset = pos as u32;
